requirements.txt
Unidecode~=1.3.6
requests~=2.28.1
librosa~=0.8.1
numba~=0.55.2
numpy==1.22.0
scipy~=1.8.1
torch~=1.13.0
jamo~=0.4.1
pypinyin~=0.47.1
jieba~=0.42.1
cn2an~=0.5.19
inflect~=5.6.2
opencc~=1.1.1
yiri-mirai
